mm: Decrease swappiness to 10
Even though we shifted to lz4, we still do not want to swap that much given that it takes CPU time for compressing and decompressing RAM data. Thus, we should only rely to it when memory pressure is high and capacity is running low. 6 or 8 gigabytes of RAM should be enough for most usage without utilizing ZRAM swap that much.

Signed-off-by: ederekun <sedrickvince@gmail.com>